Refactor URL utilities
